Created Addressing component and using it ont addAdress and EditAddress
